requirements.txt
torch==2.5.1
torchvision
numpy<2
opencv-python
Pillow
timm
scipy
scikit-image
kornia
einops

tqdm
prettytable
tabulate

huggingface-hub>0.25
accelerate
requests
transformers


PyWavelets
thop
torchsummary
fvcore
tensorboard
